+++ src/main/scala/ai/privado/audit/DependencyReport.scala
@@ -3,7 +3,7 @@ package ai.privado.audit
 import ai.privado.audit.DependencyReport.getClass
 import ai.privado.languageEngine.java.cache.DependencyModuleCache
 import ai.privado.languageEngine.java.cache.DependencyModuleCache.RuleCategoryInfo
-import ai.privado.languageEngine.java.language.module.{NodeStarters, StepsForModule}
+import ai.privado.semantic.language.*
 import io.shiftleft.codepropertygraph.generated.Cpg
 import io.shiftleft.codepropertygraph.generated.nodes.{Dependency, ModuleDependency}
 import io.shiftleft.semanticcpg.language._
