+++ src/main/scala/ai/privado/languageEngine/default/passes/PropertyVerificationPass.scala
@@ -1,12 +1,11 @@
 package ai.privado.languageEngine.default.passes
 
 import ai.privado.cache.RuleCache
-import ai.privado.languageEngine.java.language.NodeStarters
+import ai.privado.semantic.language.*
 import ai.privado.tagger.PrivadoParallelCpgPass
 import io.shiftleft.codepropertygraph.generated.Cpg
 import io.shiftleft.codepropertygraph.generated.nodes.JavaProperty
 import overflowdb.traversal.ImplicitsTmp.toTraversalSugarExt
-import ai.privado.languageEngine.java.language.StepsForProperty
 import io.shiftleft.codepropertygraph.generated.traversal.{FileTraversalExtGen, toFileTraversalExtGen}
 
 class PropertyVerificationPass(cpg: Cpg) extends PrivadoParallelCpgPass[JavaProperty](cpg) {
