+++ src/main/scala/ai/privado/languageEngine/javascript/passes/config/JSPropertyLinkerPass.scala
@@ -1,6 +1,6 @@
 package ai.privado.languageEngine.javascript.passes.config
 
-import ai.privado.languageEngine.java.language.NodeStarters
+import ai.privado.semantic.language.*
 import ai.privado.model.InternalTag
 import ai.privado.tagger.PrivadoParallelCpgPass
 import io.shiftleft.codepropertygraph.generated.{Cpg, EdgeTypes}
