+++ src/main/scala/ai/privado/languageEngine/python/tagger/sink/PythonAPITagger.scala
@@ -24,7 +24,7 @@ package ai.privado.languageEngine.python.tagger.sink
 
 import ai.privado.cache.{AppCache, FileLinkingMetadata, RuleCache}
 import ai.privado.entrypoint.{PrivadoInput, ScanProcessor}
-import ai.privado.languageEngine.java.language.{NodeStarters, StepsForProperty}
+import ai.privado.semantic.language.*
 import ai.privado.languageEngine.java.semantic.JavaSemanticGenerator
 import ai.privado.metric.MetricHandler
 import ai.privado.model.{Constants, InternalTag, Language, NodeType, RuleInfo}
