+++ src/main/scala/ai/privado/languageEngine/ruby/passes/config/RubyPropertyLinkerPass.scala
@@ -1,6 +1,6 @@
 package ai.privado.languageEngine.ruby.passes.config
 
-import ai.privado.languageEngine.java.language.NodeStarters
+import ai.privado.semantic.language.*
 import ai.privado.tagger.PrivadoParallelCpgPass
 import io.shiftleft.codepropertygraph.generated.{Cpg, EdgeTypes}
 import io.shiftleft.codepropertygraph.generated.nodes.{AstNode, JavaProperty}
