+++ src/main/scala/ai/privado/passes/YamlLinkerPass.scala
@@ -1,6 +1,6 @@
 package ai.privado.passes
 
-import ai.privado.languageEngine.java.language.NodeStarters
+import ai.privado.semantic.language.*
 import ai.privado.tagger.PrivadoParallelCpgPass
 import io.shiftleft.codepropertygraph.generated.{Cpg, EdgeTypes}
 import io.shiftleft.codepropertygraph.generated.nodes.*
