+++ src/main/scala/ai/privado/semantic/language/package.scala
@@ -23,10 +23,12 @@
 package ai.privado.semantic
 
 import ai.privado.dataflow.Dataflow
+import ai.privado.model.Constants
 import io.shiftleft.codepropertygraph.generated.nodes.*
 import io.shiftleft.codepropertygraph.generated.{Cpg, EdgeTypes, NodeTypes}
 import io.shiftleft.semanticcpg.language.{DefaultNodeExtensionFinder, NodeExtensionFinder}
 import overflowdb.traversal.*
+import io.shiftleft.semanticcpg.language.*
 
 import scala.jdk.CollectionConverters.IteratorHasAsScala
 import scala.util.Try
@@ -100,6 +102,102 @@ package object language {
     def androidXmlPermissionNode: Traversal[AndroidXmlPermissionNode] =
       Try(cpg.graph.nodes(NodeTypes.ANDROID_XML_PERMISSION_NODE).asScala.cast[AndroidXmlPermissionNode]).toOption
         .getOrElse(Iterator.empty[AndroidXmlPermissionNode])
+  }
+
+  implicit class NodeStarters(cpg: Cpg) {
+    def property: Traversal[JavaProperty] =
+      Try(cpg.graph.nodes(NodeTypes.JAVA_PROPERTY).asScala.cast[JavaProperty]).toOption
+        .getOrElse(Iterator.empty[JavaProperty])
+  }
+
+  implicit class StepsForProperty(val trav: Traversal[JavaProperty]) extends AnyVal {
+
+    def usedAt: Traversal[CfgNode] =
+      Try(trav.out(EdgeTypes.IS_USED_AT).cast[CfgNode]).toOption.getOrElse(Iterator.empty[CfgNode])
+
+    def file: Traversal[File] = Try(trav.out(EdgeTypes.SOURCE_FILE).cast[File]).toOption.getOrElse(Iterator.empty[File])
+
+  }
+
+  implicit class NodeTravToProperty(val trav: Traversal[AstNode]) {
+    def originalProperty: Traversal[JavaProperty] =
+      Try(trav.out(EdgeTypes.ORIGINAL_PROPERTY).cast[JavaProperty]).toOption.getOrElse(Iterator.empty[JavaProperty])
+  }
+
+  implicit class NodeToProperty(val node: AstNode) {
+    def originalProperty: Option[JavaProperty] = {
+      val property = node.out(EdgeTypes.ORIGINAL_PROPERTY)
+      if (property != null && property.hasNext) {
+        val prop = property.next()
+        if (prop.isInstanceOf[JavaProperty]) {
+          return Some(prop.asInstanceOf[JavaProperty])
+        }
+      }
+      None
+    }
+
+    def originalPropertyValue: Option[String] = {
+      val property = node.out(EdgeTypes.ORIGINAL_PROPERTY)
+      if (property != null && property.hasNext) {
+        val prop = property.next()
+        if (prop.isInstanceOf[JavaProperty]) {
+          return Some(prop.asInstanceOf[JavaProperty].value)
+        }
+      }
+      None
+    }
+  }
+
+  implicit class NodeToOriginalSource(val node: AstNode) extends AnyVal {
+    def originalSource: Option[AstNode] = {
+      val _originalSource = node.out(EdgeTypes.ORIGINAL_SOURCE)
+      if (_originalSource.nonEmpty && _originalSource.hasNext) {
+        return Option(_originalSource.next().asInstanceOf[AstNode])
+      }
+      None
+    }
+
+    def originalSource(sourceId: String): Option[AstNode] = {
+      val _originalSource = node.out(EdgeTypes.ORIGINAL_SOURCE)
+      if (_originalSource.nonEmpty && _originalSource.hasNext) {
+        return _originalSource
+          .find(node => node.asInstanceOf[AstNode].tag.nameExact(Constants.id).value(sourceId).nonEmpty)
+          .asInstanceOf[Option[AstNode]]
+      }
+      None
+    }
+  }
+
+  implicit class OriginalToDerivedSource(val node: AstNode) extends AnyVal {
+    def derivedSource: Option[AstNode] = {
+      val _derivedSource = node.out(EdgeTypes.DERIVED_SOURCE)
+      if (_derivedSource.nonEmpty && _derivedSource.hasNext) {
+        return Option(_derivedSource.next().asInstanceOf[AstNode])
+      }
+      None
+    }
+  }
+
+  implicit class NodeStartersForModule(cpg: Cpg) {
+
+    def module: Traversal[io.shiftleft.codepropertygraph.generated.nodes.Module] =
+      cpg.graph.nodes(NodeTypes.MODULE).asScala.cast[io.shiftleft.codepropertygraph.generated.nodes.Module]
+  }
+
+  implicit class StepsForModule(val trav: Traversal[io.shiftleft.codepropertygraph.generated.nodes.Module])
+      extends AnyVal {
+
+    def file: Traversal[File] = trav.out(EdgeTypes.SOURCE_FILE).cast[File]
+
+    def dependencies: Traversal[ModuleDependency] = trav.out(EdgeTypes.DEPENDENCIES).cast[ModuleDependency]
+  }
+
+  implicit class StepsForDependency(
+    val traversal: Traversal[io.shiftleft.codepropertygraph.generated.nodes.ModuleDependency]
+  ) extends AnyVal {
+
+    def file: Traversal[File] = traversal.out(EdgeTypes.SOURCE_FILE).cast[File]
+
   }
 
 }
