+++ src/test/scala/ai/privado/languageEngine/ruby/config/RubyEnvPropertiesFilePassTest.scala
@@ -1,8 +1,8 @@
 package ai.privado.languageEngine.ruby.config
 
 import ai.privado.cache.RuleCache
-import ai.privado.languageEngine.java.language.NodeStarters
-import ai.privado.languageEngine.java.language.*
+import ai.privado.semantic.NodeStarters
+import ai.privado.semantic.*
 import ai.privado.languageEngine.ruby.passes.config.RubyPropertyLinkerPass
 import ai.privado.model.Language
 import ai.privado.testfixtures.RubyFrontendTestSuite
